Move ShowUrlsTransform to sphinx.builders.latex.transforms
